Sync header docs and checker
