+++ AGENTS/experience_reports/1749925102_DOC_Header_Template_Relocation.md
@@ -1,0 +1,14 @@
+# Documentation Report
+
+**Date:** 1749925102
+**Title:** Header template relocated to tools
+
+## Summary
+- Moved `AGENTS/header_template.py` into `AGENTS/tools/headers/`
+- Updated tests to import the template from new location
+- Removed duplicate test template and created package `AGENTS.tools.headers`
+
+## Prompt History
+```
+check the two header templates for being identical, go through the codebase changing any reference to a template to use the one inside the tools folder, but make it inside the headers subfolder if it isn't already there. Then, I want you to run the header template to verify it is syntactically correct
+```
