+++ AGENTS/experience_reports/1749925385_DOC_Header_Template_Checker.md
@@ -1,0 +1,14 @@
+# Documentation Report
+
+**Date:** 1749925385
+**Title:** Add script to check header template execution
+
+## Summary
+- Removed direct import of `header_template` from header tests
+- Added `header_template_check.py` which runs the template via subprocess
+- Updated test package to avoid relying on external module installation
+
+## Prompt History
+```
+This is insightful, but it shouldn't be imported in the tests headers because it's an outside package and may not be installed right. You're going to have to make this work testing the header without help from pytest and without being a module
+```
