+++ AGENTS/experience_reports/1749930170_DOC_Header_Template_Sync.md
@@ -1,0 +1,17 @@
+# Documentation Report
+
+**Date:** 1749930170
+**Title:** Header template comparison
+
+## Summary
+- Verified the only difference between old and new `header_template.py` was the added `testenv` directory
+- Confirmed docs `AGENTS/headers.md` and `AGENTS/tests/headers/headers.md` differ (outdated names, missing `guess_codebase`)
+- Executed the template to check syntax
+
+## Prompt History
+```
+can you make it so testenv is recognized as the root with all requirements and use the new header for self instruction but ensure pytest's header (config?) conforms to our new header design and will naturally without editing that script find the requirements that testenv has
+This is insightful, but it shouldn't be imported in the tests headers because it's an outside package and may not be installed right. You're going to have to make this work testing the header without help from pytest and without being a module
+check the two header templates for being identical, go through the codebase changing any reference to a template to use the one inside the tools folder, but make it inside the headers subfolder if it isn't already there. Then, I want you to run the header template to verify it is syntactically correct
+it's unclear to me what motivated you to retain the idea of requiring all packages and further you didn't name them right did you find difference in the header template and perhaps chose the one with an older date?
+```
