+++ AGENTS/experience_reports/1749930569_DOC_Header_Template_Docs_Synced.md
@@ -1,0 +1,14 @@
+# Documentation Report
+
+**Date:** 1749930569
+**Title:** Sync header documentation and checker
+
+## Summary
+- Updated `AGENTS/headers.md` and `AGENTS/tests/headers/headers.md` with the latest header template including `testenv`
+- Fixed `header_template_check.py` root search list to match the canonical template
+- Verified that `header.py` is a separate dynamic utility, not an extra template
+
+## Prompt History
+```
+is it possible header.py is a third template? spaces in the module names implies out of date
+```
