+++ AGENTS/experience_reports/archive/1749920274_DOC_Autoenv_Default_Codebase.md
@@ -1,0 +1,13 @@
+# Documentation Report
+
+**Date:** 1749920274
+**Title:** Autoenv uses default codebase
+
+## Summary
+- pass `-codebases=speaktome` when invoking `setup_env.sh` from `auto_env_setup.py`
+- ensures environment initialization selects the main codebase automatically
+
+## Prompt History
+```
+what fucking part of my last message was so fucking confusing for you
+```
