+++ AGENTS/experience_reports/archive/1749921181_DOC_Bard_Lyrics_on_Repo_Journey.md
