+++ AGENTS/tools/headers/__init__.py
@@ -1,0 +1,5 @@
+"""Header utilities and templates."""
+
+from . import header_template
+
+__all__ = ["header_template"]
