+++ testing/test_hub.py
@@ -9,10 +9,48 @@ try:
 except Exception:
     import os
     import sys
+    from pathlib import Path
+
+    def _find_repo_root(start: Path) -> Path:
+        current = start.resolve()
+        required = {
+            "speaktome",
+            "laplace",
+            "tensorprinting",
+            "timesync",
+            "AGENTS",
+            "fontmapper",
+            "tensors",
+            "testenv",
+        }
+        for parent in [current, *current.parents]:
+            if all((parent / name).exists() for name in required):
+                return parent
+        return current
+
+    if "ENV_SETUP_BOX" not in os.environ:
+        root = _find_repo_root(Path(__file__))
+        box = root / "ENV_SETUP_BOX.md"
+        try:
+            os.environ["ENV_SETUP_BOX"] = f"\n{box.read_text()}\n"
+        except Exception:
+            os.environ["ENV_SETUP_BOX"] = "environment not initialized"
+        print(os.environ["ENV_SETUP_BOX"])
+        sys.exit(1)
+    import subprocess
+    try:
+        root = _find_repo_root(Path(__file__))
+        subprocess.run(
+            [sys.executable, "-m", "AGENTS.tools.auto_env_setup", str(root)],
+            check=False,
+        )
+    except Exception:
+        pass
     try:
         ENV_SETUP_BOX = os.environ["ENV_SETUP_BOX"]
     except KeyError as exc:
         raise RuntimeError("environment not initialized") from exc
+    print(f"[HEADER] import failure in {__file__}")
     print(ENV_SETUP_BOX)
     sys.exit(1)
 # --- END HEADER ---
