+++ tests/test_cli.py
@@ -26,6 +26,7 @@ except Exception:
             "AGENTS",
             "fontmapper",
             "tensors",
+            "testenv",
         }
         for parent in [current, *current.parents]:
             if all((parent / name).exists() for name in required):
