+++ tests/test_semantic_spectrum.py
@@ -10,9 +10,26 @@ except Exception:
     import os
     import sys
     from pathlib import Path
-    from AGENTS.tools.path_utils import find_repo_root
+
+    def _find_repo_root(start: Path) -> Path:
+        current = start.resolve()
+        required = {
+            "speaktome",
+            "laplace",
+            "tensorprinting",
+            "timesync",
+            "AGENTS",
+            "fontmapper",
+            "tensors",
+            "testenv",
+        }
+        for parent in [current, *current.parents]:
+            if all((parent / name).exists() for name in required):
+                return parent
+        return current
+
     if "ENV_SETUP_BOX" not in os.environ:
-        root = find_repo_root(Path(__file__))
+        root = _find_repo_root(Path(__file__))
         box = root / "ENV_SETUP_BOX.md"
         try:
             os.environ["ENV_SETUP_BOX"] = f"\n{box.read_text()}\n"
@@ -22,7 +39,7 @@ except Exception:
         sys.exit(1)
     import subprocess
     try:
-        root = find_repo_root(Path(__file__))
+        root = _find_repo_root(Path(__file__))
         subprocess.run(
             [sys.executable, "-m", "AGENTS.tools.auto_env_setup", str(root)],
             check=False,
