+++ tests/test_validate_headers.py
@@ -27,6 +27,7 @@ except Exception:
             "AGENTS",
             "fontmapper",
             "tensors",
+            "testenv",
         }
         for parent in [current, *current.parents]:
             if all((parent / name).exists() for name in required):
